Worked on revising the modules to create the parses using the new pattern.
