+++ scrapedia/packers.py
@@ -7,8 +7,11 @@ Classes: DataFramePacker
 """
 
 import abc
+from functools import partial
 
 import pandas as pd
+from cachetools.keys import hashkey
+from cachetools import cachedmethod, TTLCache
 
 
 class Packer(abc.ABC):
@@ -16,6 +19,18 @@ class Packer(abc.ABC):
 
 	Methods: pack
 	"""
+	def __init__(self, cache_maxsize: int=1, cache_ttl: int=300):
+		"""Packer's constructor.
+
+		Parameters
+		----------
+		cache_maxsize: int -- maximum number of objects to be stored
+		simultaneously on the internal cache (default 1)
+		cache_ttl: int -- time to live in seconds for internal caching of
+		data (default 300)
+		"""
+		self._cache = TTLCache(maxsize=cache_maxsize, ttl=cache_ttl)
+
 	@abc.abstractmethod
 	def pack(self, models: list):
 		"""Builds a data structure out of a list of model's.
@@ -36,6 +51,14 @@ class DataFramePacker(Packer):
 
 	Methods: pack
 	"""
+	def __init__(self, cache_maxsize: int=1, cache_ttl: int=300):
+		"""DataFramePacker's constructor.
+
+		Parameters @Packer
+		"""
+		super().__init__(cache_maxsize=cache_maxsize, cache_ttl=cache_ttl)
+
+	@cachedmethod(lambda self: self._cache, key=partial(hashkey, 'storage'))
 	def pack(self, models: list) -> pd.DataFrame:
 		"""Builds a data frame out of a list of model's.
